requirements.txt
mysql-connector-python
bs4
networkx
pandas
numpy
tldextract
lxml